borg.sh:checkout: Checkout another matching branch if appropriate
If the default branch of the upstream remote branch matches exactly, use
that as before (printing "ok").  However if that is not the case, but
another remote branch matches exactly, then create a corresponding local
branch and check that out.  If there is no such branch either, fall back
to the same fallback behavior as before, resetting the current local
branch to the expected commit, as recorded for the module.

If multiple matching branches exist, it is undefined which is used.
Configure the remote matching branch as the push target.  Use the
default branch as the upstream.

Closes #176.

+++ borg.sh
@@ -209,9 +209,36 @@ checkout () {
         else
             echo "Checkout $path ($hash)"
             echo "HEAD was $(git log --no-walk --format='%h %s' HEAD)"
-            if ! git reset --hard "$hash"
+
+            target=
+            for r in $(git branch -r --points-at=$hash | grep -v HEAD)
+            do
+                if [ "$(echo $r | cut -d/ -f2)" != "$branch" ]
+                then
+                    target=$r
+                    break
+                fi
+            done
+
+            if [ -n "$target" ]
             then
-                echo "Checkout of '$hash' into submodule path '$path' failed"
+                remote=$(echo $target | cut -d/ -f1)
+                branch=$(echo $target | cut -d/ -f2)
+                git branch --no-track $branch $target
+                git branch -u $(git symbolic-ref --short HEAD) $branch
+                if [ "$remote" != "$(git config remote.pushDefault)" ]
+                then
+                    git config remote.$branch.pushRemote $remote
+                fi
+                if ! git checkout $branch
+                then
+                    echo "Checkout of '$branch' into submodule path '$path' failed"
+                fi
+            else
+                if ! git reset --hard "$hash"
+                then
+                    echo "Checkout of '$hash' into submodule path '$path' failed"
+                fi
             fi
         fi
     fi
